requirements.txt
lpips==0.1.4
numpy==1.24.4
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
Pillow==9.3.0
scikit-image==0.19.3
scipy==1.10.1
torch>=1.12.0
torchvision>=0.10.0
six==1.16.0
tabulate==0.9.0
tensorboard==2.14.0